+++ dependencies/SFML-2.6.0/src/SFML/Audio/SoundFileWriterOgg.hpp
@@ -93,7 +93,7 @@ public:
 private:
 
     ////////////////////////////////////////////////////////////
-    /// \brief Flush blocks produced by the ogg stream, if any
+    /// \brief FlushToFile blocks produced by the ogg stream, if any
     ///
     ////////////////////////////////////////////////////////////
     void flushBlocks();