+++ dependencies/boost_1_80_0/libs/beast/test/extern/zlib-1.2.11/contrib/ada/zlib-streams.adb
@@ -112,7 +112,7 @@ package body ZLib.Streams is
 
          exit when Last < Buffer'Last;
       end loop;
-   end Flush;
+   end FlushToFile;
 
    -------------
    -- Is_Open --