+++ dependencies/boost_1_80_0/libs/beast/CHANGELOG.md
@@ -3092,7 +3092,7 @@ Version 61:
 * Tidy up some integer conversion warnings
 * Add message::header_part()
 * Tidy up names in error categories
-* Flush the output stream in the example
+* FlushToFile the output stream in the example
 * Clean close in Secure WebSocket client
 * Add server-framework SSL HTTP and WebSocket ports
 * Fix shadowing warnings